mjolnir: Allow window cycling
This cycles the other way around than cmd-< but this is the only way
that's doable easily since I just cycle the list the last window gets
the focus and is inserted at the top of the list, the rest shifts one
place down. If I'd focus the second window, it would become the first
and the first would become the second and I'd never reach the third.
This could only be solvable by storing the window list and cycling and
then when would I throw the stored list away?

The given fn is called on every switch. I just use it for Mail for now,
and this is a bit annoying since the selected mailbox also changes when
I'm switching away from the main window.