hammerspoon: Switch over from mjolnir (WIP)
Hammerspoon is integrated and it is actively maintained whereas mjolnir
uses external luarocks which is more open and more lightweight but
always makes updating a pain since you have to reinstall all rocks. And
mjolnirs author has a track record of abandoning his projects which he
already kind of did with mjolnir, proclaiming that he now works on a
new, paid version that does something similar but different and lala...
Hammerspoon seems to be the safer bet here.