git: Switch branches back after failed deletion
When deleting a branch and switching to another, the branch change
should be rolled back when the deletion fails. This means you'll be
where you started off again.

This enables you to use git cod and then just use git codd if the
current branch deletion failed because it is unmerged and you still want
to delete it.