vim: Add applescript to open nvim in new iterm win
This only handles absolute path args and just one of them. It's used
from an automator application that just takes it's input and passes it
to this script. The purpose is to be able to drop a file on the
application or assign the application as default application to certain
file types (txt, md, ...) and have those open in new terminal windows
using nvim.