key_remap_4_mac_book: Use right cmd as ctrl key
The right cmd key is much easier to reach with the right thumb than the
ctrl key (only on the left on my notebook keyboard) is with the left
pinky. And you still have all your fingers available for typing the key
combo letter. Overall, it's much easier (meaning less finger travel) and
quicker to use ctrl key combos.