Add installation script
This allows to install everything needed to use these dotfiles. It also
provides a script (dotfiles) to install/uninstall packages including the
dependencies (eg brew formulas, managed via Brewfiles) and additional install
scripts when needed.

All this is done for the zsh package as an example. This is the basic package
that's required for using these dotfiles, they are zsh based.