Nest customer packages (aka archive them)
With the new way of installing packages via the dotfiles executable, it's
possible to nest the folders/packages. So this is a wrapper (starting with a _)
that wraps alike packages. I just didn't want to delete those (although blau
and qype are no longer customers of mine), since they might contain some gems
for further use.