oh-my-zsh: Add Brewfile and change installation
Before, it was done with a submodule. Now, it's just cloned explicitly. The
installation script mainly just checks the environment and modifies rc files
which I don't want anyway. And I want to get rid of the git submodules. This
comes with the cost that I can't pin the exact version of oh-my-zsh, but I can
live with that.