sublime_text: Wipe up
I don't use sublime anymore. It was very nice. But in the end, I couldn't stand
it's closed sourceness and it's maintainer's way of dealing with bug reports
and feature requests. And it's lack of documentation. And the hassle of
figuring out, what's ST2 and what's ST3. Well, that paved my way to vim. Now,
I'll just clean up and remove my sublime text configuration.