tmux: Use $HOME/.config for config files
Up to now, I just had the version specific files in .config, now all of them are. The tmux developers have clearly stated on a github issue that they won't support XDG config file locations. This might change in the future, but apparently not in the near future. So doing this using an alias is the way to go for now.