karabiner/hammerspoon: Switch ESC/CTRL-b
Hitting Alt for CTRL-b and then hjkl for moving between panes was a pain
on the right hand. And it was a change for my muscle memory. So tapping
on Caps Lock now again sends CTRL-b as before. The right ALT key is now
my ESC key. Let's see how that works out in practice. But I only have
one key change to get used to.

CTRL-b is now also only sent when in iTerm and when tmux is actually
running to avoid outputting ^B into a terminal session that's not
running tmux.