ssh: Reenable keychain access on Sierra
On pre Sierra macOS, ssh always asked (via GUI dialog) whether to store
an ssh key passphrase in the keychain. The passphrases were then stored
in the standard login keychain.

As of Sierra, this has been disabled. You need to enable it manually.
And for using this inside of tmux, you need ot reattach the ssh process
to the user namespace or otherwise it won't have access to the
keychain [1]. The passphrases are encrypted and stored in a separate
keychain in ~/Library/Keychains/*/keychain-2.db.

[1] Here's the error output for ssh -vvv anyhost:
    debug3: Search for item with query: {
        acct = "/Users/thomas/.ssh/id_rsa";
        agrp = "com.apple.ssh.passphrases";
        class = genp;
        labl = "SSH: /Users/thomas/.ssh/id_rsa";
        nleg = 1;
        "r_Data" = 1;
        svce = OpenSSH;
    }
    debug2: Unexpected keychain error while searching for an item: Kein Schlüsselbund verfügbar. Möglicherweise musst du deinen Computer neu starten.
    Enter passphrase for key '/Users/thomas/.ssh/id_rsa':