hammerspoon: Make grid more predictable
Before, when a window refused to be positioned exactly on a grid (eg it
had minimum/maximum window sizes), the grid code wouldn't recognize it
being on that grid position on the next grid.move and would just start
over again instead of continuing over the specified grids.

Now, the current grid index is stored and iterated on.

This apparently also works better with multiple screens, where before it
could happen that you grid.move a window on one screen, move it to
another and then further grid.move calls would move it back to the
previous screen. This doesn't happen anymore.