karabiner: Fix common typo alt-space
This usually happens, when you pipe something on the shell. First, type
the pipe (alt-7) and then hit the space bar before you released the alt
key. It looks good, but the command isn't found. And you're confused,
until you see the additional space preceding the command that's not
found in the error message.

This fixes this. You cannot type a non breaking space with this, but I
don't usually need this.