git: Make paged output less annoying
It seems that git uses the pager more and more with each release. Even
short branch lists are paged, which requires an additional keystroke to
quit less.

This change tells less to output everything to stdout and quit if the
data fits on one screen. It only pages if there's a reason to page.