hammerspoon: Hide app when single window
When you press the shortcut to start/activate an app while the app is
active and the current window is the only one, then just hide the app
again.

This helps with apps like skype (that only have a single window) to
quickly show the app, check if there's something new and then hide the
window again. Just tap the shortcut twice.