karabiner/hammerspoon: Move old hyper handling
This is better done in Karabiner directly, since there are fewer layers
involved.

These hotkeys are handled by the applications directly. They obviously
can't access the state of the caps lock key or use it as a modifier, so
we need to restore the old "all modifiers at once" kind of hyper key for
these.