karabiner: Switch to generated configuration
karabiner.json gets way too bloated, especially with the kind of
wildcard tmux mappings. Generating the file is even the officially
recommended way.

The changes to karabiner.json are merely cosmetic (formatting), so the
generation does a pretty decent job for now. So this is the baseline it
can evolve on.