karabiner: Reduce timeout for shift brackets
I got parenthesis too often when accidentally hitting shift w/o needing
them and then releasing the key again. With the timeout, the tap and
release have to be quite quick but the time is sufficient to manage to
be quicker when intently tapping the shift keys for
parenthesis/brackets.