karabiner: Fix shift brackets
The curly brackets didn't work in non-terminals because the right
command key is modified to right control in terminals only. I need to
select and provide both rules based on the frontmost app.

Also, using command shift arrow keys for selecting until beginning/end
of line didn't work anymore. This is now handled by specifying the
correct modifiers and resending them in `to`.