kitty: Try hiding the mouse
This was a major problem I didn't continue switching kitty a few months
ago: The mouse pointer disappearing completely when moving it over kitty
from time to time. That's why I set this to 0 in the beginning. Trying
it out now since clicking on a word to edit it and then having it hidden
under the pointer was annoying. Maybe it works this time.