zsh/ssh: Ensure german (UTF-8) language settings
The original problem was that on some servers, the LANG environment
variable wasn't set properly and umlauts didn't work due to that (eg in
strings in a ruby REPL). First setting the LANG on ssh connections did
this but I thought it's better to make sure the local ZSH also has that
setting and just forward it. In case I might have another one locally at
some point, this will be reflected on my remote connections
automatically.