vim: Reactivate indent lines
Yggdroot/indentLine had a problem, I guess (can't remember), so I
deactivated it and replaced it with `set foldcolumn=4` to get a hint of
the indentation level left of the line number column.

I just found lukas-reineke/indent-blankline.nvim today which seems to
work quite fine. It's neovim only and written in lua. Seems to be OK,
quick and work without problems for now - right out of the box.