hammerspoon: Add shortcuts to open new windows
Opening a new browser or terminal window was sometimes a PITA. When on
another space, I always had to focus Firefox first, which switched to
another space, then open a new window and move it to the other space.
Now, it's a simple shortcut. Same with a terminal. Don't want to bring
that full screen terminal to the foreground when all I want is create a
new one.

Shift is the new default for new windows now.

- without modifier: launch/focus
- with shift: new window
- alt: alternative version (eg Tor instead of Firefox)