zsh: Use .zshenv and ZDOTDIR
There's only one file in the home directory now: .zshenv

And it only sets the environment variable ZDOTDIR to point to
~/.config/zsh. So everything else is in there:

- .zshenv (that sources other *.zshenv files)
- .zshrc (that sources other *.zsh files and configures the login shell)
- .zsh_history
- .zcompdump-* files

So zsh won't pollute ~ anymore.

And I switched to using zshenv files to configure the shell even for
non-login shells. This should prevent some of the "not found in path"
errors when starting a new window with a command (for that's
non-interactive and won't load .zshrc).