tmux/git/tig: Use default colors
git/tig just uses the default color scheme now. I had to set this
explicitly before because the terminal was light but tigs should be
dark.

tmux just uses the background color of the color schenme used in kitty
(Argonaut) now. Not the foreground color yet, since it's rather bright.