kitty: Allow launch by bundle ID and use for kitty
Strangely, `hs.application.get("kitty")` found a firefox window with
the text "kitty" in the title. That was a window object, not an
application object, not having the method isFrontmost. So that called
failed then. This might be a bug (`hs.application.get` returning a
window), but it's better to use the bundle ID anyway. apps.lua now can
handle both name and bundle ID.