zsh: Implement upgrade.d/maintain.d extension points
With this, each package can provide it's own scripts to update/maintain
itself.

Has to move some stuff from `.zshrc` to `.zshenv` because the former is
only available in interactive shells. Will have to clean quite a lot up
in the future to make sure only things relevant to interactive shells
are in `.zshrc` files.