taskwarrior: Add dirtasks for local project tasks
I want to have to possibility to track and manage project specific
tasks. These may be playground projects or something that is important
for some time but then isn't anymore. I don't want to pollute my global
tasks with these. So this ~/bin/dirtasks checks for a directory
`.dirtasks` in the local directory and uses this as taskwarrior data
directory. This is the default behaviour for `task` now, also for the
newly created alias `t`.