hammerpoon: Add browser handling
This can be used by setting Hammerspoon as default browser in the OS.
Then, each URL will be passed into the httpCallback which then can
choose which browser to open based on predefined criteria. In the first
version, only the host will be examined and other (eg local) config
files can add mappings.