zsh: Remove double load prevention for zshenv
On the second time in kitty/tmux, it didn't iterate over the .zshenv
files but all the stuff from the old iteration was gone. There must be
different processes involved. So now, it goes through the .zshenv files
twice, but at least it has the stuff defined there available all the
time in all terminals.