stats: Make force copy of preferences and make private
Before, the file was always symlinked (as per default settings in
dotfiles), but stats always (tried to) overwrite that file. In the past,
this broke the symlink. Now, it seems the write didn't succeed, so the
settings were lost on each restart, presenting me with the screen to
configure the app.

Now, the file is copied, stats can modify it properly and the change
will be displayed by `dotfiles diff` (binary change) and can be added to
the vault.